Mo 19 Sep

+++ src/main/java/at/kaindorf/jpaintro/pojos/Address.java
@@ -24,5 +24,5 @@ public class Address implements Serializable {
     @Column(nullable = false)
     private String number;
     @Column(name = "zip_code", nullable = false)
-    private Short zip;
+    private Integer zip;
 }
